Refactor migration files to ensure proper foreign key constraints removal and improve error handling in seeders

+++ database/migrations/2023_01_23_075959_add_uuid_field_to_failed_jobs_table.php
@@ -27,8 +27,11 @@ return new class extends Migration {
      */
     public function down()
     {
-        Schema::table('failed_jobs', function (Blueprint $table) {
-            $table->dropColumn('uuid');
-        });
+        if (Schema::hasTable('failed_jobs') and !Schema::hasColumn('failed_jobs', 'uuid')) {
+            Schema::table('failed_jobs', function (Blueprint $table) {
+                $table->dropColumn('uuid');
+            });
+        }
+
     }
 };
